Use ComponentWrapperHelper to simplify expander
I think it's simpler and more idiomatic to use the
GovukPublishingComponents::Presenters::ComponentWrapperHelper to build
the top-level element. Also it means the data attributes on the
top-level element can now easily be customised.

I'm planning to move the setting of the ga4-related data attributes out
of this template and this change will facilitate that.

For the moment I haven't used ComponentWrapperHelper#add_class to add
the "app-c-expander" CSS class to the top-level element, because the
validation only allows the "gem-c-*" prefix; not the "app-c-*" prefix. I
plan to address this once the component is extracted into
GovukPublishingComponents which will then make it possible to customise
the CSS classes on the top-level element.