Set option-select ga4 data attrs outside component
I want to extract the option-select component into the
govuk_publishing_components gem so that we can use it in the signon
application without duplicating the code.

The signon app doesn't yet fully support GA4 and we don't have any
immediate need to add GA tracking to our instances of option-select.
Even if we did, the tracking added here feels specific to
finder-frontend and so not suitable for extraction into the shared gem.

Setting the GA4-related data attributes outside the component decouples
the component from finder-frontend without losing any of the existing
GA4 tracking.

Also the e.g. the `ga4-change-category` data attribute is set in other
facet partials rather than in individual components, so this brings the
option-select facet into line with those.