Render each facet individually vs as a collection
This is effectively a more verbose version of the code it replaces. I
want to pass another local variable to each facet partial and this
change will make doing that easier.

The `option_select_facet_counter` local variable was previously being
provided automatically by Rails; now we're providing it explicitly.

Previously Rails was providing a `*_counter` local variable to each
facet partial where the prefix was based on the facet partial name. Now
we're providing the *same* local variable to each facet partial and,
although none of the other facet partials were using the counter, it
makes sense to give it a more generic name - hence I've named it
`index`.

There _might_ be a more elegant way to do this, but I couldn't find it!

The new version still relies on `FilterableFacet#to_partial_path`.

As an aside I found it a bit confusing that this index is 0-indexed
whereas the related index in the
GOVUK.analyticsGa4.Ga4FinderTracker.setFilterIndexes() JS function [1]
is 1-indexed. However, I haven't changed that here.

[1]: https://github.com/alphagov/finder-frontend/blob/665e954f26da3aa288f4e516b0272eb07c7ead55/app/assets/javascripts/analytics-ga4/ga4-finder-tracker.js#L10-L25