Reduce duplication in FindersController & facet_collection
The facet_collection partial only ever uses the filterable facets, so we
can avoid some duplication by selecting the filterable facets in
FindersController#facets. However, there are a few callsites within the
controller which use the full list of facets, so I've introduced a new
`#all_facets` method to handle that without making it available as a
view helper method.