WIP: Add ga4-expandable data attr in relevant components
This data attribute doesn't seem to depend on anything else so it's
simpler to move it into the relevant components at the cost of a small
amount of duplication. This is a small step towards making
addFilterButtonTracking() redundant.

Note that I may move the setting of this data attribute out into the
facet partials in a subsequent commit.