WIP: Set ga4-event & ga4-expandable in facet partials
As opposed to in option-select.js or expander.js thereby avoiding the
need for GOVUK.analyticsGa4.Ga4FinderTracker.setFilterIndexes() to
trigger the "ga4-filter-indexes-added" custom event and the need to call
GOVUK.analyticsGa4.Ga4FinderTracker.addFilterButtonTracking() from
option-select.js & expander.js.

Now the ga4-related data attributes to be attached to the button created
dynamically by the components are supplied in the
`button_data_attributes` local variable in the relevant facet partial.
This is more awkward than I would've liked, because of the need to make
the button data attributes available at component JS initialization time
as opposed to just at component partial render time.

The awkwardness described above makes me think it would be worth
investigating some kind of middle ground where _some_ of the GA4
functionality is optionally enabled within the components but the
relevant distinguishing variables are still set in the facet partials,
i.e. it's probably OK for the components to "know" that they should set
`event_name` to "select_content" in the "ga4-event" data attribute, but
it's probably not OK for the components to "know" that they should set
`type` to "finder", so the latter should be passed in to the component
from the facet partial.

TODO:

* Make equivalent changes to the expander component and all the facet
  partials which directly or indirectly use it. This may be a bit more
  awkward!
* Consider making the index from FacetsIterator 1-indexed rather than
  0-indexed so we don't have to keep adding 1 in several places.
* Consier reducing duplication in the generation of the data attributes
  in the facet partials; perhaps by extracting a helper method.
* Completely remove setFilterIndexes() & addFilterButtonTracking()
* Do we still need to set the ga4-index data attribute on the component
  top-level element or was this only used by the component JS to obtain
  the ga4-index data? It would simplify the code in the facet partials if
  we only needed to pass the ga4-index data in the
  `button_data_attributes` and not in the `data_attributes` as well.
* Update docs/analytics-ga4/ga4-finder-tracker.md &
  docs/analytics-ga4/ga4-filter-expand-collapse.md docs to reflect the
  new approach.