Pass AB params to Search::Query
The ab_params is only needed to determine which version of search-api
to call. It will not need be needed in search-api itself.
Therefore rather than updating all of the tests to stub passing in the
`ab_params` the `ab_params` argument is no longer being passed to the
QueryBuilder. The `ab_params` are not being used in QueryBuilder at the
moment, so there won't be a noticeable change to the behaviour.

The tests uses "anything" as an RSpec matcher. This is a positional
argument that allows you specify "anything" if you don't care what the
other arguments are.