Use ab_params to call search-api-v2
Uses `ab_params` that are passed from the controller to the search
query to determine which version of search-api to use.

The query_param "use_v2" will still be available to use the new
version of search.

The controller tests need to be updated to stub both search-apis as
we're testing all three AB test variants.