Only show parts for top three results
There is a requirement for only the top three search results to show
parts, which is currently handled by `search-api` only returning parts
for the top three results. However, this should really be a frontend
concern, and shouldn't be handled by the backend returning different
result data depending on what the position of a result is across all
results.

- Determine the `result_number` for each result in `ResultSetPresenter`
  and pass on to `SearchResultPresenter` (ideally this would use "rank"
  for better semantics but that's currently used by a debug parameter)
- Determine in `SearchResultPresenter` whether or not to include parts
  in the presentation depending on the result number