Get Google attribution token from search api v2
The discovery engine attribution token is the unique id for every query
set to Google Vertex Search, the new search provider for site search.

Every time a ticket is logged with Google support about search results
the first thing they ask for is the attribution token for the request.

This attribution token is being returned in the response from
search-api-v2, but as the api is not exposed on the internet it is
difficult to get the token without logging into Google Cloud Platform.

Instead the attribution token should be surfaced in search results in
the user interface along with other debugging information so that
anyone can access it.

The token is defaulted to nil so that the search v1 results are unaffected.
This can be removed whenever V1 id decommissioned.