Display Google attribution token on search results
The discovery engine attribution token is the unique id for every query
set to Google Vertex Search, the new search provider for site search.

Every time a ticket is logged with Google support about search results
the first thing they ask for is the attribution token for the request.

This attribution token is being returned in the response from
search-api-v2, but as the api is not exposed on the internet it is
difficult to get the token without logging into Google Cloud Platform.

Instead the attribution token should be surfaced in search results in
the user interface along with other debugging information so that
anyone can access it.

If debug_score is true, display the Google discovery engine attribution
token.

The attribution token is the unique id for the whole query, so it only
needs to be displayed once.

It is hidden behind the debug_score to match the hidden elastic search
scores.

The styling also matches elastic search scores.