Use `search-api` v1 for non-keyword searches
We're moving away from our initial intent to use Search API v2
exclusively for all queries to "site search" (aka `/search/all`) for the
short to medium term.

This is for two reasons:
- Our underlying search engine technology isn't designed (yet) to handle
  non-relevance based queries well (resulting in e.g. attempting to view
  _all_ documents matching certain filter criteria sorted by updated
  missing many documents)
- We get a lot of queries without keywords from automated processes,
  which costs money in a usage-based charging model

In the longer term, this should be resolved by a more flexible content
API that allows retrieving doucments by certain criteria. Until then, we
can continue to use v1 as it is still around and we know it's able to
serve non-relevance based search scenarios well.