Support inverse style on topic pages that have a query string
The requirement to render an inverse blue-banner full-width version of topic pages was set out in https://trello.com/c/A6eidvxn/799-implement-blue-topic-page-finder-headers-on-supergroup-finders-when-user-has-come-from-topic-page-m. An example of such a page is here: https://www.gov.uk/search/guidance-and-regulation?topic=c58fdadd-7743-46d6-9629-90bb3ccc4ef0. I asked around to see if we still need to retain this inverse design seeing that it seem to have been related to the migration of some pages. However, we don't seem to have enough information to definitely say that this inverse blue banner version of topic pages could be retired so I'm retaining it as part of this work I'm doing to move breadcrumbs ouf to the main container.

This commit adds styles and a wrapper class for the phase banner and the breadcrumbs to set a full-width blue background. It simplifies the syntax of checking if this inverse design is required by adding a helper for this. It uses a wrapper class to set the inverse before-content styles.