Filter panel: Give panel its own bottom border
The filter panel is always supposed to have a bottom border, but
currently relies on the top border of the document list following it for
its border.

This gives the panel its own border so it's present even if there is no
document list following it, and accordingly removes it from the top of
the first document list item on the new all content finder view.