Add `SortPresenter#default?`
This returns whether or not the results are sorted by their default sort
order. The logic to determine this is slighly nasty as it has to take
some edge cases into account (that seem to be possible according to
other specs), in particular the situation where the Finder does not
specify a default sort option.