New all content finder UI: Add ability to sort
This adds `SortFacet`, which acts as a "virtual" facet, in the sense
that it's not based on a facet that comes through from the finder
content item. It implements the minimum method signature to duck type as
a `Facet`.

The new "all content" finder UI treats sorting as part of the overall
filter UI, instead of as a separate section on the page. Having a facet
object for sorting allows us to reuse a lot of the existing logic around
rendering facets, in particular the way that they are iterated over and
rendered on to the page, and how their GA4 tracking is generated.