All content finder: Implement hidden clearable facet
This implements the hidden clearable facet for the new "all content"
finder UI. These are facets that are able to be used with the finder as
query parameters, and if used are shown to the user, but do not have UI
to manually select. We can have a much simpler implementation than the
existing finder UI because we do not need to consider Javascript.