Introduce `FacetPresenter`
The current logic for dealing with section indices and counts is quite
messy, with the `FacetsIterator` yielding several values that are then
passed through several layers of views and partials.

This introduces a `FacetPresenter` that holds section index/count values
(appropriately using 1-based indexing) and makes the iterator yield that
instead.