All content finder: Fix tracking of filter removal
The filter "tags" when removed include the visually hidden prefix which
is messing up analytics. This adds an explicit `text` property to the
event equal to only the actually displayed text.

Also refactors the displayed text into a property on the item returned
by the presenter for the filter tags.