DateInput: Allow converting to hash or string
As we want both the old UI to be able to use query parameters intended
for the new UI, and vice versa, for example when the user has a link to
a search bookmarked, we need to be able to convert between the two for
display purposes (to show the value in the form field).

This adds a `#to_str` and `#to_hash` method to `DateInput` to achieve
this, which we will use in a future commit to display the values in the
respective form fields.