DateValidator: Ignore empty fields in hash dates
There was a situation where `DateValidator` was checking whether the raw
input was blank (it wasn't as it had keys, but all keys were blank), and
assuming that the fact that `DateParser` returned `nil` meant the date
didn't validate (rather than just being empty).

This ensures if the date input is a Hash it will get blank fields
compacted before being checked for presence.