`DateHashParser`: Expect positive date values
We don't want to be too strict in validating, but we should never expect
a non-positive value for any date field. This could be because the user
has entered zero or a negative value, or because `#to_i` couldn't figure
out the input string and returned its zero sentinel value.