Refactor `DateStringParser`
This class is very messy and hard to follow. In particular, it has an
instance method `#parse` that takes an argument, which is then passed on
to countless instance methods.

This refactors it to keep the date string as an instance variable set on
initialization, and also updates the logic to be much simpler and easier
to follow.