GA4 finder tracker: Handle new three-field date facets
Our new design uses the canonical GOV.UK Design System three-field date
input. This needs special handling and can't just be treated as date
fields.

The logic agreed is as follows:
- A select event is fired on any change of the three date fields,
  provided that after the change all three fields are filled in
- A remove event is fired on a change after which all three fields are
  empty
- Any change after which 1 or 2 fields are filled in/empty will not
  result in an event
- The event's text will be the contents of the three fields, in order,
  joined by slashes, e.g. "13/12/1989"
- The event's section will be the overall name of the field, e.g. "Date
  before"

This requires the introduction of a new "date" type for the GA4 finder
tracker and wrapping the date fields in a container div with the right
section information (as the `date_input` component doesn't support
arbitrary HTML element data attributes).