`DateHashParser`: Be less lenient with invalid values
The current logic in `DateHashParser` discards invalid (non-positive or
string) values, which can be confusing as the user's input is partially
accepted without triggering a validation failure.

This makes the parser less lenient, and bails out on parsing if any
present part of the date hash isn't a positive number.