All content finder: Properly track empty keywords
GOV.UK Analytics's `standardiseSearchTerm` returns undefined for empty
queries, whereas for the `search` event we are sending, we _do_ want to
keep track of the search term even if a blank string, and using
undefined there will not update the state properly.

Because this adds a default value for the keyword form field, we need to
adjust the test for the empty field submission logic to consider the
keyword present.