All content finder: Use short name for date fields
This changes the date facet and filter tags for the all content finder
to use the "short name" instead of the full name for the facet fields.
Specifically, this is because we want to use the regular name ("Date")
for the section title, but the short name ("Updated") for the fields
within.

It also updates the feature test suite fixture for the all content
finder with the most recent version of the live content item and adjusts
the Cucumber features and steps accordingly.