Add govuk_chat_promo_helper
We want to show a promo to the new GOV.UK Chat application on certain pages
on GOV.UK.

This helper adds a list of base paths we want to render the chat promo
banner on. For now this is just the business finance support page, but
more may be added in the future.

If the GOVUK_CHAT_PROMO_ENABLED env var is set to true and the base
path passed in matches a base path from the constant the #show_govuk_chat_promo?
method returns true.