Display GOV.UK Chat promo on certain pages
We want to show a promo to the new GOV.UK Chat application on certain pages
on GOV.UK.

We already have this implemented in Government Frontend
https://github.com/alphagov/government-frontend/pull/3303
and Collections https://github.com/alphagov/collections/pull/3742.

We're implementing it in:

- finder_frontend
- smart_answers
- frontend

This updates the finder show page to render the chat banner above the
related links similar to the other applications.