Exclude top margin on intervention on inverse taxon page
We have a different design for taxon pages when they have the topic query string eg. /search/guidance-and-regulation?topic=c58fdadd-7743-46d6-9629-90bb3ccc4ef0 Exclude the top margin we're adding on the intervention banner in this inverse design view as it results in a blank gap in the full width header.(NB: This taxon inverse design got added for some historical page migration but we're not confident we can remove it. However, the fact that it's not clear it exists when developers change stuff is probably another reason to consider removing it now...)