Add a GOV.UK Chat banner to search results
This adds the GOV.UK Chat promo banner to be shown on the right hand
side of the GOV.UK Search results page. It uses an env var defined in
GOV.UK Helm Charts [1].

It will be shown on every GOV.UK search results page. It is expected to
be removed on December 20th - first by env var and then later by code
removal.

This design and wording has been through various people for approval:

- GOV.UK AI design
- GOV.UK AI Lead PM
- GOV.UK Director

[1]: https://github.com/alphagov/govuk-helm-charts/pull/2837