Derive metadata labels from allowed values on facet where present
For many document types, the values indexed in Elasticsearch for its metadata fields are the kebab case value from the select option value in the Publishing app.

These values are not suitable for presenting on GOV.UK, so there is a configuration in each content type defined in Search API called `expanded_search_result_fields`. This configuration is used to replace the kebab case value with a hash containing the original value and the human friendly label it relates to in the Search API query results.

Maintaining this configuration is painful, particularly for specialist document types, because it means that every time a department wants to add, change or remove an option on a facet, we have to update the configuration in both the Publishing tool and in Search API, and we have to keep them in sync.

However, given the label provided in the expanded search result fields configuration never varies from the label present in the `allowed_values` attribute on the facet configuration, we may as well fetch the label from there and remove the `expanded_search_result_fields` from Search API, which will reduce the effort required from Publishing teams to maintain Specialist Finders.

This commit extracts the appropriate metadata labels from the `allowed_values` configuration on the facet if one is provided. It will work both for search results that have had the field expansion applied, and those that haven't, so should be safe to release before the removal of the configuration. It also preserves the existing behaviour for facets that do not set `allowed_values`.