Generate custom NestedFacets from raw facets
The goal with nested facets is to render two somewhat custom selects. The "child"/sub facet select would be aware of its "parent"/main facet options, thus enabling the behaviour whereby upon selecting the main, the subfacet select gets populated with the options corresponding to the main facet only (when JS is enabled).

We are sending the data almost ready to use to the content-store, the "allowed_values" being set to the relevant "label" and "value" for each facet, respectively. We send some additional fields such as the `nested_facet` boolean which is needed by the FacetBuilder when building new facets.  If a `nested_facet` field is present, we generate a custom NestedFacet.

Once the NestedFacet is instantiated, the only difference between it being a main or a sub facet is stored in some instance variables. We need to know which facet is the "main", i.e. the one bearing the `sub_facet_key`.

The nested facet's hash will additionally have references to the parent values, to help us render the correct select options and labels, and enable the js behaviour. See future commits.