Extract facet options for NestedFacets
The facet hash that a NestedFacet is initialised with, will include ready to use allowed values - i.e. an array of {"label": "", "value": ""} objects - and some additional fields that allow us to distinguish whether a facet is main or sub.

The corresponding view will render a select for the nested facet, so we're mapping the data from label/value to text/value, which is what a select component requires.

Each allowed value of the facet, if the facet is a subfacet (i.e. not a main), additionally stores a reference to its parent/main facet. This comes through as a `main_facet_value` field from the content store. We've updates Specialist Publisher presenters to send this, and extended the publishing api schema to allow it. The `main_facet_value` is then mapped to a data-attribute that will appear in the rendered HTML. This will be consumed by the JS (see future commits). The main facet options do not require a data-attribute.