Add javascript to render nested facet values
The logic in the js ensures that:
- when an "all" main facet option is selected, the sub-facet is disabled;
- only the relevant "sub" facet options are showing for a certain "main" facet selection;
- the "all" sub facet options are default upon selecting a main option.

The `_nested_facet` view gets rendered for each of the facets. For a "nested facet behaviour" to be accomplished we'd require at least two facets, a main and a sub facet. Nonetheless, we wouldn't want the js to be loaded twice, nor is it desirable to load it by default. So we're only loading it once for the main facet. In order to reach the subfacet select, we're using a document level selector in the js.

In order to customize the js to work with any facet and subfacet keys, we're additionally sending along the keys as data attributes to the module.