Add separator to the bottom of the subfacet select
Other finders' facet sidebars typically receive a separator at the bottom, which seems to come for free from the components they use. The select does not render one, so we're adding it manually. Nonetheless, we only want to add the separator at the bottom of the "sub" facets, not the "main".

Also added a bit of margin/padding where needed.