Introduce label data attribute to render subfacet when js is disabled
Whilst the raw facet hash used to build the NestedFacet includes ready to use labels, when JS is not enabled, the subfacet select renders ALL the options, regardless of the main facet selection.

In order to cater for this, we're going to render the subfacet labels in a dasherized fashion by default: "Main facet label - Sub facet label", so it is easier for the users to distinguish between label duplicates, such as "Not applicable" (when js is disabled).

To compose the labels, we make use of the `main_facet_label`, which is now set for each raw subfacet allowed value in content store (we've updates Specialist Publisher presenters to send this, and extended the publishing api schema to allow it).

When the js loads, it will drop the main facet label from the child, as the subfacet select will now only render the options relevant to the selected main facet value so the user can easily identify the desired select option.