Remodel NestedFacet to behave more like a TaxonFacet
We want to make the newly added `NestedFacet` a bit more consistent with the pre-existent, and somewhat similar, `TaxonFacet`. There are two behaviour points that are ultimately the goal:
- getting the applied filters/metadata tags at the top of the search results to behave similarly to taxon, i.e. visually represent that there is a logical unit of a nested facet rather than two distinct facets
- wrapping the nested facet in an expander/accordion behaviour, similar to how other facets behave.

In the previous implementation, the `NestedFacet` was written up as two distinct facets of type "text", with some additional fields to keep track of which is the "main" and which is the "sub". The scope of that change was a lot smaller than if we implemented a custom facet, with most of the changes in the originating finder repo, "Specialist Publisher".

The view that served each facet did not lend itself to the accordion behaviour, so we looked into making a single nested facet again. Taxon is a good example to follow, as it, too, has two selects/facets under the hood, but behaves like a "single" facet.

Nonetheless, `TaxonFacet` populates its options in a unique way, whilst `NestedFacet` will get its options from Content Store, and ultimately from the schema defined in Specialist Publisher. There are also differences in how the two are implemented in Search API. For `NestedFacet`, we have opted to maintain two facets both in Search API and Publishing API schemas, whilst `TaxonFacet` seemingly has its own custom rules in Search API, and likely send a unified query.

Additional changes will be made in subsequent commits. This commit now lays out the main `NestedFacet` class and tests. The approach is to build a single `NestedFacet` from the raw "nested" data facet we get from Content Store, and in doing so, replicate desired behaviours from `TaxonFacet`, and follow its implementation as far as we can. To this end, the new implementation uses a new "type" - "nested" - to make it easier to carve out the custom logic further down the line.

Whilst further refactoring is possible - abstract out common behaviours between the two classes, or use inheritance - I believe it is too early to make that call. Further steps in the `NestedFacets` work will include expanding the selection to multiple options, which will diverge from taxon behaviour. The test coverage of nested behaviours in this and upcoming commits should also support this future behaviour expansion.

These changes are accompanied by corresponding changes in Specialist Publisher and Publishing API.