Allow overriding serving config in `QueryBuilder`
We need to be able to specify different serving configs for a query
handled by Search API v2 based on AB test values. Unfortunately Finder
Frontend needs this to be passed quite deep down the call stack.

This amends `QueryBuilder` to take a `v2_serving_config` parameter that
defaults to the default serving config, but can be overridden to a
different one if necessary.