Add custom filter for building the nested facet query
We need to send both main and sub facet values in the query. This is because in our implementation, a "nested" facet does eventually translate into two distinct tags for main and sub facet values, respectively, in Search API and PublishingAPI (i.e. the documents are tagges with two distinct tags).

Although we are trying to model based on the taxon behaviour, it appears that the `TaxonFilter` sends a single key (with a tuple for value), which doesn't suit our needs. Nor does any other Filter, such as Text etc.