Render applied filters
This commit replicates a bunch of public methods from the `TaxonFacet` into the `NestedFacet`. These methods are used by the `FiltersPresenter` to render the metadata tags at the top of the search results.

The gist of the change is that, despite a `NestedFacet` being a single logical unit, we want to render metadata tags/applied filters for both main and sub facet. The resulting UI consists of an "and"-ed sentence fragment that reads as `<Main facet> <Main facet selected value> \n and <Sub Facet><Sub facet selected value>`. The two facets are rendered on the same background colour, rather than alternating (as is the case with two distinct facets).

The `NestedFacet` behaviour is now consistent with the `TaxonFacet`.